Consolidated floatingTextManager into floatingText.py
